refactor(reporting): move pages to views/reporting/pages.py
Split the guide + builder page routes (reporting_guide, reporting,
_guide_render, _GUIDE_MD, _ai_effort_enabled) out of
nx_lib/views/reporting/__init__.py into pages.py, part of beautify-phase-2a
Task 3. __init__.py re-exports _GUIDE_MD/_guide_render and delegates
register_routes to pages.register_routes(app).

Retargeted the has_permission patch in
test_reporting_page_caption_slots_need_only_explain_data to
nx_lib.views.reporting.pages.has_permission.
